Autherisation Logic added for all resources with testing

+++ backend/src/main/java/ch/hftm/blog/boundry/BlogLikeResource.java
@@ -8,15 +8,19 @@ import ch.hftm.blog.entity.User;
 import ch.hftm.blog.service.BlogLikeService;
 import ch.hftm.blog.service.BlogService;
 import ch.hftm.blog.service.UserService;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.SecurityContext;
 import org.eclipse.microprofile.openapi.annotations.Operation;
 
 @Path("blog/like")
 public class BlogLikeResource {
+
     @Inject
     BlogLikeService blogLikeService;
 
@@ -27,28 +31,41 @@ public class BlogLikeResource {
     UserService userService;
 
     @POST
+    @RolesAllowed({"admin", "user"})
     @Path("/{userId}/{blogId}")
     @Produces(MediaType.APPLICATION_JSON)
+    @Consumes(MediaType.APPLICATION_JSON)
     @Transactional
     @Operation(summary = "Add like to blog by a user")
-    public Response addLikeToBlog(@PathParam("userId") long userId,
+    public Response addLikeToBlog(@Context SecurityContext securityContext,
+                                  @PathParam("userId") long userId,
                                   @PathParam("blogId") long blogId) {
         Response.Status status = Response.Status.OK;
         Object dto = null;
         boolean isSuccess = true;
+
         try {
-            User user = userService.getUserById(userId);
-            Blog blog = blogService.findBlogById(blogId);
-            if (user == null) {
-                status = Response.Status.NOT_FOUND;
-                dto = new ErrorResponseDTO1("User not found");
-                isSuccess = false;
-            } else if (blog == null){
-                status = Response.Status.NOT_FOUND;
-                dto = new ErrorResponseDTO1("Blog not found");
+            String loggedInUserId = securityContext.getUserPrincipal().getName();
+            long loggedInUserIdLong = Long.parseLong(loggedInUserId);
+
+            if (userId != loggedInUserIdLong) {
+                status = Response.Status.FORBIDDEN;
+                dto = new ErrorResponseDTO1("User is not authorized to like this blog");
                 isSuccess = false;
             } else {
-                blogLikeService.addLikeToBlog(blog, user);
+                User user = userService.getUserById(userId);
+                Blog blog = blogService.findBlogById(blogId);
+                if (user == null) {
+                    status = Response.Status.NOT_FOUND;
+                    dto = new ErrorResponseDTO1("User not found");
+                    isSuccess = false;
+                } else if (blog == null) {
+                    status = Response.Status.NOT_FOUND;
+                    dto = new ErrorResponseDTO1("Blog not found");
+                    isSuccess = false;
+                } else {
+                    blogLikeService.addLikeToBlog(blog, user);
+                }
             }
         } catch (Exception e) {
             status = Response.Status.INTERNAL_SERVER_ERROR;
@@ -60,28 +77,47 @@ public class BlogLikeResource {
 
     @DELETE
     @Path("/{userId}/{blogId}")
+    @RolesAllowed({"admin", "user"})
+    @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     @Transactional
     @Operation(summary = "Remove like from blog by a user")
-    public Response removeLikeToBlog(@PathParam("userId") Long userId,
+    public Response removeLikeToBlog(@Context SecurityContext securityContext,
+                                     @PathParam("userId") Long userId,
                                      @PathParam("blogId") Long blogId) {
         Response.Status status = Response.Status.OK;
         Object dto = null;
         boolean isSuccess = true;
+
         try {
-            User user = userService.getUserById(userId);
-            Blog blog = blogService.findBlogById(blogId);
-            if (user == null) {
-                status = Response.Status.NOT_FOUND;
-                dto = new ErrorResponseDTO1("User not found");
-                isSuccess = false;
-            } else if (blog == null){
-                status = Response.Status.NOT_FOUND;
-                dto = new ErrorResponseDTO1("Blog not found");
+            String loggedInUserId = securityContext.getUserPrincipal().getName();;
+            long loggedInUserIdLong = Long.parseLong(loggedInUserId);
+
+            if (userId != loggedInUserIdLong) {
+                status = Response.Status.FORBIDDEN;
+                dto = new ErrorResponseDTO1("User is not authorized to remove like from this blog");
                 isSuccess = false;
             } else {
-                BlogLike blogLike = blogLikeService.getBlogLikeByBlogAndUserID(blogId, userId);
-                blogLikeService.removeLikeToBlog(blogLike);
+                User user = userService.getUserById(userId);
+                Blog blog = blogService.findBlogById(blogId);
+                if (user == null) {
+                    status = Response.Status.NOT_FOUND;
+                    dto = new ErrorResponseDTO1("User not found");
+                    isSuccess = false;
+                } else if (blog == null) {
+                    status = Response.Status.NOT_FOUND;
+                    dto = new ErrorResponseDTO1("Blog not found");
+                    isSuccess = false;
+                } else {
+                    BlogLike blogLike = blogLikeService.getBlogLikeByBlogAndUserID(blogId, userId);
+                    if (blogLike == null) {
+                        status = Response.Status.NOT_FOUND;
+                        dto = new ErrorResponseDTO1("Like not found");
+                        isSuccess = false;
+                    } else {
+                        blogLikeService.removeLikeToBlog(blogLike);
+                    }
+                }
             }
         } catch (Exception e) {
             status = Response.Status.INTERNAL_SERVER_ERROR;
